+++ dags/stg/infra_logs/truncate_old.py
